move core.get_generators to generation.get_generators
