+++ src/logic/scripting/lua/libcore.cpp
@@ -4,7 +4,6 @@
 #include "constants.hpp"
 #include "engine.hpp"
 #include "content/Content.hpp"
-#include "content/ContentLoader.hpp"
 #include "files/engine_paths.hpp"
 #include "files/settings_io.hpp"
 #include "frontend/menu.hpp"
@@ -180,32 +179,6 @@ static int l_get_default_generator(lua::State* L) {
     return lua::pushstring(L, WorldGenerator::DEFAULT);
 }
 
-/// @brief Get a list of all world generators
-/// @return A table with the IDs of all world generators
-static int l_get_generators(lua::State* L) {
-    const auto& packs = engine->getContentPacks();
-
-    lua::createtable(L, 0, 0);
-
-    int i = 0;
-    auto names = ContentLoader::scanContent(
-        ContentPack::createCore(engine->getPaths()), ContentType::GENERATOR);
-    for (const auto& name : names) {
-        lua::pushstring(L, name);
-        lua::rawseti(L, i + 1);
-        i++;
-    }
-    for (const auto& pack : packs) {
-        auto names = ContentLoader::scanContent(pack, ContentType::GENERATOR);
-        for (const auto& name : names) {
-            lua::pushstring(L, name);
-            lua::rawseti(L, i + 1);
-            i++;
-        }
-    }
-    return 1;
-}
-
 const luaL_Reg corelib[] = {
     {"new_world", lua::wrap<l_new_world>},
     {"open_world", lua::wrap<l_open_world>},
@@ -219,5 +192,4 @@ const luaL_Reg corelib[] = {
     {"get_setting_info", lua::wrap<l_get_setting_info>},
     {"quit", lua::wrap<l_quit>},
     {"get_default_generator", lua::wrap<l_get_default_generator>},
-    {"get_generators", lua::wrap<l_get_generators>},
     {NULL, NULL}};
